Revert "Added srcId field"
This reverts commit bc5718d47715212ab2326a13dc2a0d351ecfa475.

+++ outbound-sdk/src/main/java/com/aerospike/connect/outbound/BinData.java
@@ -55,7 +55,7 @@ public class BinData {
      * The src-id of XDR.
      */
     @Nullable
-    private final Short srcId;
+    private final Byte srcId;
 
     /**
      * The meta info of the list if a {@link #value} is a list.
@@ -79,7 +79,7 @@ public class BinData {
     }
 
     public BinData(@Nonnull ParticleType particleType, @Nullable Object value,
-                   @Nullable Long lastUpdateTimeMs, @Nullable Short srcId) {
+                   @Nullable Long lastUpdateTimeMs, @Nullable Byte srcId) {
         this(particleType, value, lastUpdateTimeMs, srcId, null, null);
     }
 }
